Move precompiled Groovy plugins coss version tests to pluginDevelopment submodule
