+++ subprojects/integ-test/integ-test.gradle.kts
@@ -46,8 +46,6 @@ dependencies {
     crossVersionTestImplementation(project(":ide"))
     crossVersionTestImplementation(project(":codeQuality"))
     crossVersionTestImplementation(project(":signing"))
-    crossVersionTestImplementation(project(":logging"))
-
 
     allTestRuntimeDependencies.forEach {
         crossVersionTestRuntimeOnly(it)
